refactor: Group button components
